SocketIO architecture
- Extended Architecture.md, with a quick simple explanation on how it works.
- Exported event name constants.
- Moved all previous event logic in Index[main].ts to the new architecture style.
- Renamed Join.sio.ts to UserJoin.sio.ts
- Updated JSDoc for each new event file.
