+++ server/src/socketio-events/scrum-poker/events/UserJoin.sio.ts
@@ -3,6 +3,7 @@ import UserHandler from "../../../handlers/UserHandler";
 import IUser from "../../../interfaces/IUser";
 import User from "../../../dto/User.dto";
 import socketio, { Socket } from "socket.io";
+import { USER_JOIN } from "../constants/EVENT_CONSTANTS";
 
 /**
  * The primary event for joining a room with a predetermined estimate
@@ -10,9 +11,9 @@ import socketio, { Socket } from "socket.io";
  * @class
  * @event Join#constructor:join
  */
-export default class Join extends SocketIOEvent {
+export default class UserJoin extends SocketIOEvent {
     constructor(io: socketio.Server , socket: Socket) {
-        super("join", ({ users_name, room, estimate }: IUser, acknowledgeFn) => {
+        super(USER_JOIN, ({ users_name, room, estimate }: IUser, acknowledgeFn) => {
             UserHandler.addUserToLocalStore(new User(
                 users_name.trim().toLowerCase(),
                 room.trim().toLowerCase(),
