+++ server/src/socketio-events/scrum-poker/events/index.ts
@@ -1,3 +1,11 @@
-import Join from "./Join.sio";
+import UserJoin from "./UserJoin.sio";
+import UserSendEstimate from "./UserSendEstimate.sio";
+import UserViewEstimates from "./UserViewEstimates.sio";
+import UserDisconnect from "./UserDisconnect.sio";
 
-export default [Join];
+export default [
+    UserJoin,
+    UserSendEstimate,
+    UserViewEstimates,
+    UserDisconnect
+];
